connectingDB/workingWithDB/pyproject.toml
[project]
name = "workingwithdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.5.0",
    "python-dotenv>=1.2.1",
]
